requirements.txt
aiohttp
aiosqlite
discord.py
python-dotenv
spacy==3.8.2
matplotlib
ollama